backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
google-generativeai==0.3.2
faster-whisper==1.0.3
python-dotenv==1.0.0
aiofiles==23.2.1
